wip: Move status functions out of timeline cases in to statusrepository
TimelineCases is increasingly badly named, as some of these methods
operate on statuses irrespective of whether or not they're on a
timeline.

Move reblog, favourite, and bookmark methods in to a new
StatusRepository.
