+++ core/data/src/main/kotlin/app/pachli/core/data/repository/notifications/NotificationsRepository.kt
@@ -22,7 +22,6 @@ import androidx.paging.InvalidatingPagingSourceFactory
 import androidx.paging.Pager
 import androidx.paging.PagingConfig
 import androidx.paging.PagingData
-import app.pachli.core.common.PachliError
 import app.pachli.core.common.di.ApplicationScope
 import app.pachli.core.data.model.StatusViewData
 import app.pachli.core.database.dao.NotificationDao
@@ -37,20 +36,11 @@ import app.pachli.core.database.model.NotificationEntity
 import app.pachli.core.database.model.RemoteKeyEntity
 import app.pachli.core.database.model.RemoteKeyEntity.RemoteKeyKind
 import app.pachli.core.database.model.StatusViewDataEntity
-import app.pachli.core.eventhub.BookmarkEvent
 import app.pachli.core.eventhub.EventHub
-import app.pachli.core.eventhub.FavoriteEvent
-import app.pachli.core.eventhub.PollVoteEvent
-import app.pachli.core.eventhub.ReblogEvent
 import app.pachli.core.model.AccountFilterDecision
 import app.pachli.core.model.FilterAction
 import app.pachli.core.network.model.Notification
 import app.pachli.core.network.retrofit.MastodonApi
-import app.pachli.core.network.retrofit.apiresult.ApiError
-import com.github.michaelbull.result.Err
-import com.github.michaelbull.result.Ok
-import com.github.michaelbull.result.Result
-import com.github.michaelbull.result.onFailure
 import com.github.michaelbull.result.onSuccess
 import javax.inject.Inject
 import kotlinx.coroutines.CoroutineScope
@@ -58,30 +48,6 @@ import kotlinx.coroutines.async
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.launch
 import timber.log.Timber
-
-/**
- * Errors that can occur acting on a status.
- *
- * Converts the [throwable] from a
- * [NetworkResult][at.connyduck.calladapter.networkresult.NetworkResult] to
- * a [PachliError].
- *
- * @param throwable The wrapped throwable.
- */
-// TODO: The API calls should return an ApiResult, then this can wrap those.
-sealed class StatusActionError(open val error: ApiError) : PachliError by error {
-    /** Bookmarking a status failed. */
-    data class Bookmark(override val error: ApiError) : StatusActionError(error)
-
-    /** Favouriting a status failed. */
-    data class Favourite(override val error: ApiError) : StatusActionError(error)
-
-    /** Reblogging a status failed. */
-    data class Reblog(override val error: ApiError) : StatusActionError(error)
-
-    /** Voting in a poll failed. */
-    data class VoteInPoll(override val error: ApiError) : StatusActionError(error)
-}
 
 /**
  * Repository for [NotificationData] interacting with the remote [MastodonApi]
@@ -237,140 +203,6 @@ class NotificationsRepository @Inject constructor(
     fun setExpanded(pachliAccountId: Long, statusViewData: StatusViewData, isExpanded: Boolean) = externalScope.launch {
         saveStatusViewData(pachliAccountId, statusViewData.copy(isExpanded = isExpanded))
     }
-
-    /**
-     * Sets the bookmark state of [statusId] to [bookmarked].
-     *
-     * Sends [BookmarkEvent] if successful.
-     *
-     * @param pachliAccountId
-     * @param statusId ID of the status to affect.
-     * @param bookmarked New bookmark state.
-     */
-    suspend fun bookmark(
-        pachliAccountId: Long,
-        statusId: String,
-        bookmarked: Boolean,
-    ): Result<Unit, StatusActionError.Bookmark> = externalScope.async {
-        val deferred = async {
-            if (bookmarked) {
-                mastodonApi.bookmarkStatus(statusId)
-            } else {
-                mastodonApi.unbookmarkStatus(statusId)
-            }
-        }
-
-        statusDao.setBookmarked(pachliAccountId, statusId, bookmarked)
-
-        val result = deferred.await()
-
-        result.onFailure { throwable ->
-            statusDao.setBookmarked(pachliAccountId, statusId, !bookmarked)
-            return@async Err(StatusActionError.Bookmark(throwable))
-        }
-
-        result.onSuccess { eventHub.dispatch(BookmarkEvent(statusId, bookmarked)) }
-
-        return@async Ok(Unit)
-    }.await()
-
-    /**
-     * Sets the favourite state of [statusId] to [favourited].
-     *
-     * Sends [FavoriteEvent] if successful.
-     *
-     * @param pachliAccountId
-     * @param statusId ID of the status to affect.
-     * @param favourited New favourite state.
-     */
-    suspend fun favourite(
-        pachliAccountId: Long,
-        statusId: String,
-        favourited: Boolean,
-    ): Result<Unit, StatusActionError.Favourite> = externalScope.async {
-        val deferred = async {
-            if (favourited) {
-                mastodonApi.favouriteStatus(statusId)
-            } else {
-                mastodonApi.unfavouriteStatus(statusId)
-            }
-        }
-
-        statusDao.setFavourited(pachliAccountId, statusId, favourited)
-
-        val result = deferred.await()
-
-        result.onFailure { throwable ->
-            statusDao.setFavourited(pachliAccountId, statusId, !favourited)
-            return@async Err(StatusActionError.Favourite(throwable))
-        }
-
-        result.onSuccess { eventHub.dispatch(FavoriteEvent(statusId, favourited)) }
-
-        return@async Ok(Unit)
-    }.await()
-
-    /**
-     * Sets the reblog state of [statusId] to [reblogged].
-     *
-     * Sends [ReblogEvent] if successful.
-     *
-     * @param pachliAccountId
-     * @param statusId ID of the status to affect.
-     * @param reblogged New bookmark state.
-     */
-    suspend fun reblog(
-        pachliAccountId: Long,
-        statusId: String,
-        reblogged: Boolean,
-    ): Result<Unit, StatusActionError.Reblog> = externalScope.async {
-        val deferred = async {
-            if (reblogged) {
-                mastodonApi.reblogStatus(statusId)
-            } else {
-                mastodonApi.unreblogStatus(statusId)
-            }
-        }
-
-        statusDao.setReblogged(pachliAccountId, statusId, reblogged)
-
-        val result = deferred.await()
-
-        result.onFailure { throwable ->
-            statusDao.setReblogged(pachliAccountId, statusId, !reblogged)
-            return@async Err(StatusActionError.Reblog(throwable))
-        }
-
-        result.onSuccess { eventHub.dispatch(ReblogEvent(statusId, reblogged)) }
-
-        return@async Ok(Unit)
-    }.await()
-
-    /**
-     * Votes [choices] in [pollId] in [statusId],
-     *
-     * Sends [PollVoteEvent] if successful.
-     *
-     * @param pachliAccountId
-     * @param statusId ID of the status to affect.
-     * @param pollId ID of the poll to vote in.
-     * @param choices Array of indices of the poll options being voted for.
-     */
-    suspend fun voteInPoll(
-        pachliAccountId: Long,
-        statusId: String,
-        pollId: String,
-        choices: List<Int>,
-    ): Result<Unit, StatusActionError.VoteInPoll> = externalScope.async {
-        mastodonApi.voteInPoll(pollId, choices)
-            .onSuccess { poll ->
-                statusDao.setVoted(pachliAccountId, statusId, poll.body)
-                eventHub.dispatch(PollVoteEvent(statusId, poll.body))
-            }
-            .onFailure { return@async Err(StatusActionError.VoteInPoll(it)) }
-
-        return@async Ok(Unit)
-    }.await()
 
     companion object {
         private const val PAGE_SIZE = 30
